mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "fastmcp-test>=0.1.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy>=2.0.44",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
]
